+++ test/spec/features/context-pad/ContextPadProvider.js
@@ -5,13 +5,17 @@ import imageC from './resources/c.png';
 import { every } from 'min-dash';
 
 
-export default function ContextPadProvider(contextPad) {
+export default function ContextPadProvider(contextPad, elementRegistry) {
+  this._contextPad = contextPad;
+  this._elementRegistry = elementRegistry;
+
   contextPad.registerProvider(this);
 }
 
-ContextPadProvider.$inject = [ 'contextPad' ];
+ContextPadProvider.$inject = [ 'contextPad', 'elementRegistry' ];
 
 ContextPadProvider.prototype.getContextPadEntries = function(element) {
+  var self = this;
 
   if (element.type === 'A') {
     return {
@@ -28,8 +32,7 @@ ContextPadProvider.prototype.getContextPadEntries = function(element) {
         }
       }
     };
-  } else
-  if (element.type === 'drag') {
+  } else if (element.type === 'drag') {
     return {
       'action.dragstart': {
         className: 'drag-out',
@@ -38,6 +41,24 @@ ContextPadProvider.prototype.getContextPadEntries = function(element) {
             e.__handled = true;
 
             return 'action.dragstart';
+          }
+        }
+      }
+    };
+  } else if (element.type === 'hover') {
+    return {
+      'action.hover': {
+        className: 'hover',
+        action: {
+          click: function(e) {
+            self._contextPad.open(self._elementRegistry.get('s2'));
+
+            return 'action.click';
+          },
+          hover: function(e) {
+            e.__handled = true;
+
+            return 'action.hover';
           }
         }
       }